feat(parser): notes can have more than one paragraph
test/expected file names change to .json.

BREAKING CHANGE: The regex for matching notes are loosen. The semicolon and space are optional. The `notes` object is no longer a key-value object but an array of note object, such as
```js
{
  title: 'BREAKING AMEND',
  text: 'some breaking change'
}
```
The detection of notes, closes, continueNote and isBody are mutually exclusive.

Closes #3